+++ lib/src/lazy_stream.dart
@@ -4,8 +4,9 @@
 
 import "dart:async";
 
-import "stream_completer.dart";
 import "delegate/stream.dart";
+import "stream_completer.dart";
+import "utils.dart";
 
 /// A [Stream] wrapper that forwards to another [Stream] that's initialized
 /// lazily.
@@ -15,13 +16,11 @@ import "delegate/stream.dart";
 /// produce a `Stream`.
 class LazyStream<T> extends Stream<T> {
   /// The callback that's called to create the inner stream.
-  ZoneCallback _callback;
+  FutureOrCallback<Stream<T>> _callback;
 
   /// Creates a single-subscription `Stream` that calls [callback] when it gets
   /// a listener and forwards to the returned stream.
-  ///
-  /// The [callback] may return a `Stream` or a `Future<Stream>`.
-  LazyStream(callback()) : _callback = callback {
+  LazyStream(FutureOr<Stream<T>> callback()) : _callback = callback {
     // Explicitly check for null because we null out [_callback] internally.
     if (_callback == null) throw new ArgumentError.notNull('callback');
   }
@@ -41,7 +40,7 @@ class LazyStream<T> extends Stream<T> {
     var result = callback();
 
     Stream<T> stream;
-    if (result is Future) {
+    if (result is Future<Stream<T>>) {
       stream = StreamCompleter.fromFuture(result.then((stream) {
         return DelegatingStream.typed/*<T>*/(stream as Stream);
       }));